refactor(doctor): move roster scope helpers
